+++ tests/testthat/test-plugins.R
@@ -40,7 +40,7 @@ crew_test("custom launcher", {
     reset_packages = FALSE,
     reset_options = FALSE,
     garbage_collection = FALSE,
-    launch_max = 5L
+    crashes_error = 5L
   ) {
     client <- crew::crew_client(
       name = name,
@@ -64,7 +64,7 @@ crew_test("custom launcher", {
       reset_packages = reset_packages,
       reset_options = reset_options,
       garbage_collection = garbage_collection,
-      launch_max = launch_max,
+      crashes_error = crashes_error,
       tls = tls
     )
     controller <- crew::crew_controller(
@@ -146,7 +146,7 @@ crew_test("custom launcher with local async errors", {
     reset_packages = FALSE,
     reset_options = FALSE,
     garbage_collection = FALSE,
-    launch_max = 5L,
+    crashes_error = 5L,
     processes = NULL
   ) {
     client <- crew::crew_client(
@@ -171,7 +171,7 @@ crew_test("custom launcher with local async errors", {
       reset_packages = reset_packages,
       reset_options = reset_options,
       garbage_collection = garbage_collection,
-      launch_max = launch_max,
+      crashes_error = crashes_error,
       tls = tls,
       processes = processes
     )
@@ -282,7 +282,7 @@ crew_test("custom launcher with async internal launcher tasks", {
     reset_packages = FALSE,
     reset_options = FALSE,
     garbage_collection = FALSE,
-    launch_max = 5L,
+    crashes_error = 5L,
     processes = NULL
   ) {
     client <- crew::crew_client(
@@ -307,7 +307,7 @@ crew_test("custom launcher with async internal launcher tasks", {
       reset_packages = reset_packages,
       reset_options = reset_options,
       garbage_collection = garbage_collection,
-      launch_max = launch_max,
+      crashes_error = crashes_error,
       tls = tls,
       processes = processes
     )