+++ NEWS.md
@@ -1,4 +1,10 @@
-# crew 0.10.1 (development)
+# crew 0.10.2
+
+* Eliminate spurious `launch_max` error from underutilized workers (#189).
+* Deprecate `launch_max` in favor of `crashes_error` (#189).
+* Look for crashes of all workers in `rotate()` instead of looking for crashes of a specific worker in `launch()` (#189).
+
+# crew 0.10.1
 
 * Instrument `crew_eval()` with `autometric::log_phase_set()` and `autometric::log_phase_reset()`.